完成Schedule模块的数据类，并将User拆分为Teacher和Student

+++ courseji/src/main/kotlin/com/littlecorgi/courseji/course/controller/CourseController.kt
@@ -62,7 +62,7 @@ class CourseController {
     @ApiOperation(value = "删除课程")
     @GetMapping(path = ["/delete"])
     fun deleteCourse(
-        @ApiParam(value = "要被删除的课程的id", required = true) @RequestParam id: Int
+        @ApiParam(value = "要被删除的课程的id", required = true) @RequestParam id: Long
     ): ServerResponse<String> {
         return try {
             ServerResponse.createBySuccess(courseService.deleteCourse(id))
